Extracted With-methods to their own class
